Update API Address
# Conflicts:
#	pom.xml
#	src/main/java/Spring_AdamStore/controller/AddressController.java
#	src/main/java/Spring_AdamStore/dto/basic/EntityBasic.java
#	src/main/java/Spring_AdamStore/entity/Address.java
#	src/main/java/Spring_AdamStore/entity/District.java
#	src/main/java/Spring_AdamStore/entity/Province.java
#	src/main/java/Spring_AdamStore/entity/Ward.java
#	src/main/java/Spring_AdamStore/mapper/AddressMapper.java
#	src/main/java/Spring_AdamStore/schedule/OrderStatusScheduler.java
#	src/main/java/Spring_AdamStore/service/impl/AddressServiceImpl.java
#	src/main/java/Spring_AdamStore/service/impl/OrderServiceImpl.java
#	src/main/resources/application-dev.yml
#	src/main/resources/db/migration/V1__create_address_tables.sql

+++ src/main/java/Spring_AdamStore/dto/request/AddressRequest.java
@@ -21,9 +21,9 @@ public class AddressRequest {
 
     @NotNull(message = "districtId không được để trống")
     @Min(value = 1, message = "districtId phải lớn hơn 0")
-    private Long districtId;
+    private Integer districtId;
 
     @NotNull(message = "provinceId không được để trống")
     @Min(value = 1, message = "provinceId phải lớn hơn 0")
-    private Long provinceId;
+    private Integer provinceId;
 }
