+++ src/main/java/Spring_AdamStore/repository/DistrictRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DistrictRepository extends JpaRepository<District, Long> {
+public interface DistrictRepository extends JpaRepository<District, Integer> {
 
     Page<District> findByProvinceId(Integer provinceId, Pageable pageable);
